web/requirements.txt
Flask==2.1.2
Flask-PyMongo
python-dotenv
requests